feat(sportify-frontend): sticky header, responsive, header navigation  add css variables (#1)
