+++ src/App.jsx
@@ -1,14 +1,15 @@
 import "./App.css";
-import Header from "./components/Header";
 import { Route, Routes } from "react-router-dom";
 import Layout from "./Layouts/Layout ";
-import Home from "./Components/Home";
+import Home from "./Routes/Home";
+import Admin from "./Routes/Admin";
 
 function App() {
   return (
     <Routes>
       <Route path="/" element={<Layout />}>
         <Route path="/" element={<Home />} />
+        <Route path="/admin" element={<Admin />} />
       </Route>
     </Routes>
   );
