+++ src/Layouts/Layout .jsx
@@ -1,14 +1,14 @@
-import React from "react";
 import { Outlet } from "react-router-dom";
 import Header from "../components/Header";
+import Footer from "../components/Footer";
 
 const Layout = () => {
   return (
-    <div>
+    <>
       <Header/>
       <Outlet />
-
-    </div>
+      <Footer/>
+    </>
   );
 };
