+++ src/Routes/Admin.jsx
@@ -1,0 +1,10 @@
+const Home = () => {
+  return (
+    <main>
+      <h2>Contenido Admin</h2>
+      <p>Aquí va la información del Admin</p>
+    </main>
+  );
+};
+
+export default Home;
