+++ src/Routes/Home.jsx
@@ -1,0 +1,12 @@
+import '../Styles/home.css';
+
+const Home = () => {
+  return (
+    <main>
+      <h2>Contenido Principal</h2>
+      <p>Aquí va la información básica de del sitio web.</p>
+    </main>
+  );
+};
+
+export default Home;
