+++ src/Styles/footer.css
@@ -1,0 +1,9 @@
+footer {
+  background: #333;
+  color: #fff;
+  text-align: center;
+  padding: 10px 0;
+  width: 100%;
+
+  height: 10%;
+}
