+++ src/Styles/home.css
@@ -1,0 +1,7 @@
+main {
+  height: 70%;
+  /* width: 100%; */
+
+  padding: 20px;
+  text-align: center;
+}
