+++ src/Styles/index.css
@@ -1,7 +1,9 @@
-#root {
-height: 100vh;
-width: 100vw;  
-}
+/* #root {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
+} */
 
 :root {
   /* Primary colores */
@@ -38,10 +40,4 @@ body {
   background-color: var(--background-color);
 
   height: 100vh;
-}
-
-a {
-  color: var(--text-primary);
-  text-decoration: none;
-  cursor: pointer;
 }
