+++ src/components/Footer.jsx
@@ -1,0 +1,11 @@
+import '../Styles/footer.css';
+
+const Footer = () => {
+  return (
+    <footer>
+      <p>2025 Sportify. Digital House.</p>
+    </footer>
+  );
+};
+
+export default Footer;
