--- src/components/Home.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-
-const Home = () => {
-  return (
-    <div>
-      <main>
-        <h2>Contenido Principal</h2>
-        <p>Aquí va la información básica de del sitio web.</p>
-      </main>
-      <footer>
-        <p>2025 Sportify. Digital House.</p>
-      </footer>
-    </div>
-  );
-};
-
-export default Home;
